Move SessionContext to authenticate as it is only used there
